WebApp/requirements.txt
tensorflow==2.5.0
streamlit==0.84.0
numpy==1.19.4
pillow==8.0.1
htbuilder
